pyproject.toml
[project]
name = "spritegrid"
version = "0.2.0"
description = "A command line tool for turning AI-generated pixel art into real pixel art."
readme = "README.md"
# requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.10.1",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "rembg>=2.0.69",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=9.0.2",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[project.scripts]
spritegrid = "spritegrid:cli.cli"
spritegrid-crop = "spritegrid:cli.crop_scale_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
testpaths = [
    "tests",
]
